requirements.txt
litellm>=1.55.0
python-dotenv>=1.0.0
pytest>=9.0.0
